+++ matchbox-frontend/src/app/validate/validation-entry.ts
@@ -1,13 +1,14 @@
 import {IssueSeverity, OperationResult} from '../util/operation-result';
 import {ValidationParameter} from "./validation-parameter";
-import { parseFhirResource } from '../util/fhir-resource-parser';
+import { FhirResource, parseFhirResource } from '../util/fhir-resource-parser';
 
 export class ValidationEntry {
   readonly filename: string; // "package/package.json",
-  readonly resource: string;
+  readonly content: string;
+  readonly resource: FhirResource;
   resourceType: string;
+  format: FhirFormat;
   resourceId: string | null;
-  readonly mimetype: string;
   result: OperationResult | undefined;
   readonly extractedProfiles: string[] = [];
   validationProfile: string | null;
@@ -22,16 +23,30 @@ export class ValidationEntry {
               settings: ValidationParameter[] = [],
               validationProfile: string | null = null) {
     this.filename = filename;
-    this.resource = resource;
+    this.content = resource;
     this.validationParameters = settings;
 
     if (mimetype) {
-      this.mimetype = mimetype;
+      switch (mimetype) {
+        case 'application/fhir+json':
+        case 'application/json':
+        case 'json':
+          this.format = FhirFormat.JSON;
+          break;
+        case 'application/fhir+xml':
+        case 'application/xml':
+        case 'text/xml':
+        case 'xml':
+          this.format = FhirFormat.XML;
+          break;
+        default:
+          throw new Error(`Unsupported mimetype ${mimetype}`);
+      }
     } else {
       if (filename.endsWith('.json')) {
-        this.mimetype = 'application/fhir+json';
+        this.format = FhirFormat.JSON;
       } else {
-        this.mimetype = 'application/fhir+xml';
+        this.format = FhirFormat.XML;
       }
     }
 
@@ -42,9 +57,14 @@ export class ValidationEntry {
     if (!parsed) {
       throw new Error(`Failed to parse resource from file ${filename}`);
     }
+    this.resource = parsed;
     this.resourceType = parsed.resourceType;
     this.resourceId = parsed.id;
     this.extractedProfiles.push(...parsed.profiles);
+  }
+
+  public get mediaType() {
+    return (this.format == FhirFormat.JSON) ? 'application/fhir+json' : 'application/fhir+xml';
   }
 
   getErrors(): number | undefined {
@@ -73,3 +93,5 @@ export class ValidationEntry {
     this.result = OperationResult.fromOperationOutcome(operationOutcome);
   }
 }
+
+export enum FhirFormat { JSON, XML}