+++ docs/changelog.md
@@ -1,4 +1,5 @@
 - Fix debug mode of the $transform operation (#572)
+- GUI: ask the engine to analyze Bundle documents for possible validation profiles (#355)
 
 2026/08/31 Release 4.1.14